trow can proxy any registry wihtout prior cfg
