+++ src/registry/api_types.rs
@@ -22,8 +22,9 @@ pub struct ReadyStatus {
 pub struct MetricsResponse {
     pub metrics: String,
 }
-/// These types are largely stripped down versions of the Kubernetes types.
-/// In future, we could directly use k8s types, but I'd rather leave that to a higher level.
+
+// These types are largely stripped down versions of the Kubernetes types.
+// In future, we could directly use k8s types, but I'd rather leave that to a higher level.
 
 pub struct ContentInfo {
     pub length: u64,
