+++ src/routes/blob.rs
@@ -1,14 +1,14 @@
 use std::sync::Arc;
 
 use axum::Router;
-use axum::extract::{Path, State};
+use axum::extract::{Path, Query, State};
 use axum::routing::get;
 
 use super::macros::endpoint_fn_7_levels;
 use crate::TrowServerState;
 use crate::registry::manifest::ManifestReference;
-use crate::registry::server::PROXY_DIR;
 use crate::registry::{BlobReader, Digest};
+use crate::routes::extracts::ImageNamespace;
 use crate::routes::macros::route_7_levels;
 use crate::routes::response::errors::Error;
 use crate::routes::response::trow_token::TrowToken;
@@ -28,26 +28,19 @@ async fn get_blob(
     _auth_user: TrowToken,
     State(state): State<Arc<TrowServerState>>,
     Path((mut repo, digest)): Path<(String, Digest)>,
+    Query(query): Query<ImageNamespace>,
 ) -> Result<BlobReader<impl tokio::io::AsyncRead>, Error> {
     let digest_str = digest.as_str();
-    if repo.starts_with(PROXY_DIR) {
-        let (proxy_cfg, image) = match state
-            .registry
-            .config
-            .registry_proxies
-            .get_proxy_config(&repo, &ManifestReference::Digest(digest.clone()))
-            .await
-        {
-            Some(cfg) => cfg,
-            None => {
-                return Err(Error::NameInvalid(format!(
-                    "No registered proxy matches {repo}"
-                )));
-            }
-        };
-        // This is important to transform f/docker.io/ubuntu into f/docker.io/_library/ubuntu
-        repo = format!("f/{}/{}", proxy_cfg.host, image.get_repo())
+    if let Some(image) = state
+        .registry
+        .config
+        .registry_proxies
+        .get_proxied_image(&repo, &ManifestReference::Digest(digest.clone()), query.ns)
+        .await
+    {
+        repo = format!("f/{}/{}", image.get_host(), image.get_repo())
     }
+
     let rowid = sqlx::query_scalar!(
         r#"
         SELECT b.rowid as "rowid!" FROM blob b
@@ -82,7 +75,8 @@ endpoint_fn_7_levels!(
     get_blob(
         auth_user: TrowToken,
         state: State<Arc<TrowServerState>>;
-        path: [image_name, digest: Digest]
+        path: [image_name, digest: Digest],
+        query: Query<ImageNamespace>
     ) -> Result<BlobReader<impl tokio::io::AsyncRead>, Error>
 );
