+++ src/routes/blob_upload.rs
@@ -402,10 +402,7 @@ mod tests {
         let range = resp_header!(resp, test_utilities::RANGE_HEADER);
         let location = resp_header!(resp, test_utilities::LOCATION_HEADER);
         assert_eq!(range, "0-0"); // Haven't uploaded anything yet
-        assert_eq!(
-            location,
-            format!("/v2/{repo_name}/blobs/uploads/{uuid}")
-        );
+        assert_eq!(location, format!("/v2/{repo_name}/blobs/uploads/{uuid}"));
 
         let blob = "super secret blob".as_bytes();
         let digest = Digest::digest_sha256_slice(blob);
@@ -440,11 +437,9 @@ mod tests {
         let resp = router
             .clone()
             .oneshot(
-                Request::post(format!(
-                    "/v2/{repo_name}/blobs/uploads/?digest={digest}"
-                ))
-                .body(Body::from(config))
-                .unwrap(),
+                Request::post(format!("/v2/{repo_name}/blobs/uploads/?digest={digest}"))
+                    .body(Body::from(config))
+                    .unwrap(),
             )
             .await
             .unwrap();
