+++ tests/smoke_test.rs
@@ -141,11 +141,13 @@ mod smoke_test {
         let data_trow1 = temp_dir.subdir_untracked("1");
 
         let trow0 = start_trow(&data_trow0, false).await;
+        let trow0_host = format!("127.0.0.1:{}", trow0.port);
         let trow1 = trow_router(&data_trow1, |cfg| {
             cfg.config_file = Some(ConfigFile {
                 registry_proxies: RegistryProxiesConfig {
                     registries: vec![SingleRegistryProxyConfig {
-                        host: format!("http://127.0.0.1:{}", trow0.port),
+                        host: trow0_host.clone(),
+                        insecure: true,
                         ..Default::default()
                     }],
                     ..Default::default()
@@ -176,7 +178,7 @@ mod smoke_test {
             .1
             .clone()
             .oneshot(
-                Request::get("/v2/f/trow/alpine/manifests/latest")
+                Request::get(format!("/v2/f/{trow0_host}/alpine/manifests/latest"))
                     .body(Body::empty())
                     .unwrap(),
             )
@@ -193,11 +195,13 @@ mod smoke_test {
         let data_trow1 = temp_dir.subdir_untracked("1");
 
         let trow0 = start_trow(&data_trow0, true).await;
+        let trow0_host = format!("[::1]:{}", trow0.port);
         let trow1 = trow_router(&data_trow1, |cfg| {
             cfg.config_file = Some(ConfigFile {
                 registry_proxies: RegistryProxiesConfig {
                     registries: vec![SingleRegistryProxyConfig {
-                        host: format!("http://[::1]:{}", trow0.port),
+                        host: trow0_host.clone(),
+                        insecure: true,
                         ..Default::default()
                     }],
                     ..Default::default()
@@ -228,7 +232,7 @@ mod smoke_test {
             .1
             .clone()
             .oneshot(
-                Request::get("/v2/f/trow/alpine/manifests/latest")
+                Request::get(format!("/v2/f/{trow0_host}/alpine/manifests/latest"))
                     .body(Body::empty())
                     .unwrap(),
             )
